rename orchard key mod to reflect contents
* I suspect the previous naming convention was inherited
* It's clear from context (the orchard keys are _exposed_ through a
wallet mod, that they're "wallet" keys in this library.
* orchard is explicit "o" is ambiguous
